src/requirements.txt
aiohttp==3.9.1 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.9" and python_version < "3.11"
attrs==23.2.0 ; python_version >= "3.9" and python_version < "4.0"
brotli==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and implementation_name == "cpython"
brotlicffi==1.1.0.0 ; python_version >= "3.9" and python_version < "4.0" and implementation_name != "cpython"
certifi==2023.11.17 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
discord-py==2.3.2 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.6 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
mutagen==1.47.0 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pycryptodomex==3.20.0 ; python_version >= "3.9" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.9" and python_version < "4.0"
yt-dlp==2023.12.30 ; python_version >= "3.9" and python_version < "4.0"
